Update previews and also fix and polish UI
